# projects/01-ai-knowledge-base/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  middleware.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/middleware.ErrorDetail'
      meta:
        $ref: '#/definitions/middleware.MetaInfo'
      success:
        type: boolean
    type: object
  middleware.MetaInfo:
    properties:
      path:
        type: string
      request_id:
        type: string
      timestamp:
        type: string
    type: object
...
  middleware.SuccessResponse:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/middleware.MetaInfo'
      success:
        type: boolean
    type: object
...
  models.KnowledgeBase:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      document_count:
        type: integer
      id:
        type: string
      is_active:
        type: boolean
      last_updated_at:
        type: string
      name:
        type: string
      owner_id:
        description: From sub-project 0
        type: string
      settings:
        $ref: '#/definitions/models.JSONMap'
      tags:
        items:
          type: string
        type: array
      total_size_bytes:
        type: integer
      updated_at:
        type: string
      visibility:
        description: public, private, shared
        type: string
    type: object
...
  service.UpdateKBRequest:
    properties:
      description:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      settings:
        $ref: '#/definitions/models.JSONMap'
      tags:
        items:
          type: string
        type: array
      visibility:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@saleschampionhub.com
    name: 技术支持
  description: "企业级AI知识库管理平台，支持多租户、RLS数据隔离和向量搜索功能。\n\n## \U0001F510 认证说明\n所有接口都需要JWT
    Token认证（从子项目0用户中心获取）。\n使用用户中心的 `/api/v1/auth/login` 接口获取Token。\n\n## ✨ 核心特性\n-
    多租户架构：租户/组织/用户三级隔离\n- Row-Level Security (RLS)：数据库级别的数据隔离\n- 三级挂载系统：租户 > 组织 >
    用户权限继承\n- 向量搜索：基于pgvector的语义搜索（1024维向量）\n- Redis多层缓存：热点数据缓存优化\n- 完整用户中心集成：统一认证授权"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: AI知识库管理平台 API
  version: "1.0"
paths:
...
  /knowledge-bases:
    get:
      description: 分页获取知识库列表，支持多种过滤条件
      parameters:
      - default: 20
        description: 每页数量（默认：20）
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量（默认：0）
        in: query
        name: offset
        type: integer
      - default: created_at DESC
        description: 排序字段（默认：created_at DESC）
        in: query
        name: order_by
        type: string
      - description: 可见性过滤（public/private/restricted）
        in: query
        name: visibility
        type: string
      - description: 所有者ID过滤
        in: query
        name: owner_id
        type: string
      - description: 搜索关键词（在名称和描述中搜索）
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 知识库列表
          schema:
            allOf:
            - $ref: '#/definitions/middleware.SuccessResponse'
            - properties:
                data:
                  properties:
                    knowledge_bases:
                      items:
                        $ref: '#/definitions/models.KnowledgeBase'
                      type: array
                    limit:
                      type: integer
                    offset:
                      type: integer
                    total:
                      type: integer
                  type: object
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 列出知识库
      tags:
      - 知识库管理
    post:
      consumes:
      - application/json
      description: 创建一个新的知识库。如果未指定所有者，将自动设置为当前用户。
      parameters:
      - description: 知识库创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateKBRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 知识库创建成功
          schema:
            allOf:
            - $ref: '#/definitions/middleware.SuccessResponse'
            - properties:
                data:
                  properties:
                    knowledge_base:
                      $ref: '#/definitions/models.KnowledgeBase'
                  type: object
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建知识库
      tags:
      - 知识库管理
  /knowledge-bases/{id}:
    delete:
      description: 软删除知识库，同时会删除所有关联的文档和向量数据
      parameters:
      - description: 知识库ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 知识库删除成功
          schema:
            allOf:
            - $ref: '#/definitions/middleware.SuccessResponse'
            - properties:
                data:
                  properties:
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: 知识库ID必填
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: 知识库不存在
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除知识库
      tags:
      - 知识库管理
    get:
      description: 获取指定知识库的详细信息，包括统计数据
      parameters:
      - description: 知识库ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 知识库详情
          schema:
            allOf:
            - $ref: '#/definitions/middleware.SuccessResponse'
            - properties:
                data:
                  properties:
                    knowledge_base:
                      $ref: '#/definitions/models.KnowledgeBase'
                    stats:
                      type: object
                  type: object
              type: object
        "400":
          description: 知识库ID必填
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: 知识库不存在
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取知识库详情
      tags:
      - 知识库管理
    put:
      consumes:
      - application/json
      description: 更新知识库的名称、描述、设置等信息
      parameters:
      - description: 知识库ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateKBRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 知识库更新成功
          schema:
            allOf:
            - $ref: '#/definitions/middleware.SuccessResponse'
            - properties:
                data:
                  properties:
                    knowledge_base:
                      $ref: '#/definitions/models.KnowledgeBase'
                  type: object
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: 知识库不存在
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新知识库
      tags:
      - 知识库管理
  /knowledge-bases/{id}/mounts:
    get:
      description: 获取指定知识库的所有挂载记录，包括租户、组织和用户级别的挂载
      parameters:
      - description: 知识库ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 挂载列表
          schema:
            allOf:
            - $ref: '#/definitions/middleware.SuccessResponse'
            - properties:
                data:
                  properties:
                    mounts:
                      items:
                        $ref: '#/definitions/models.KnowledgeBaseMount'
                      type: array
                    total:
                      type: integer
                  type: object
              type: object
        "400":
          description: 知识库ID必填
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取知识库的挂载列表
      tags:
      - 知识库挂载
  /knowledge-bases/{id}/stats:
    get:
      description: 获取知识库的详细统计信息，包括文档数量、总大小、状态分布等
      parameters:
      - description: 知识库ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 知识库统计信息
          schema:
            allOf:
            - $ref: '#/definitions/middleware.SuccessResponse'
            - properties:
                data:
                  properties:
                    document_count:
                      type: integer
                    status_breakdown:
                      type: object
                    total_size_bytes:
                      type: integer
                  type: object
              type: object
        "400":
          description: 知识库ID必填
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取知识库统计
      tags:
      - 知识库管理
  /mounts:
    get:
      description: 获取挂载列表，支持按类型、租户ID、组织ID、用户ID过滤
      parameters:
      - description: 挂载类型（tenant/organization/user）
        in: query
        name: mount_type
        type: string
      - description: 租户ID
        in: query
        name: tenant_id
        type: string
      - description: 组织ID
        in: query
        name: organization_id
        type: string
      - description: 用户ID
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 挂载列表
          schema:
            allOf:
            - $ref: '#/definitions/middleware.SuccessResponse'
            - properties:
                data:
                  properties:
                    mounts:
                      items:
                        $ref: '#/definitions/models.KnowledgeBaseMount'
                      type: array
                    total:
                      type: integer
                  type: object
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取挂载列表
      tags:
      - 知识库挂载
  /mounts/{id}:
    delete:
      description: 移除知识库挂载，撤销指定级别的访问权限
      parameters:
      - description: 挂载ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 挂载已移除
          schema:
            allOf:
            - $ref: '#/definitions/middleware.SuccessResponse'
            - properties:
                data:
                  properties:
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: 挂载ID无效
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: 挂载不存在
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 取消知识库挂载
      tags:
      - 知识库挂载
    get:
      description: 获取单个挂载记录的详细信息
      parameters:
      - description: 挂载ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 挂载详情
          schema:
            allOf:
            - $ref: '#/definitions/middleware.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.KnowledgeBaseMount'
              type: object
        "400":
          description: 挂载ID无效
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: 挂载不存在
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取挂载详情
      tags:
      - 知识库挂载
...
securityDefinitions:
  BearerAuth:
    description: 输入格式："Bearer " + JWT令牌（注意Bearer后面有空格）
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"